feat: add Cart, ProductDetail and Products pages
Add Cart page to display the products added to the cart and allow the user to
clear the cart. Add ProductDetail page to display the details of a product and
allow the user to add it to the cart. Add Products page to display all the
available products.

The pages use the useQuery hook from react-query to fetch data from the API and
display it. The pages also use the useCartStore hook from the cart store to
manage the products added to the cart. The pages use the Button and Loading
components from the libs folder to display buttons and loading spinners. The
pages use the ProductCart and ProductCard components from the libs folder to
display the products in the cart and the available products.